+++ src/hedgesnear.js
@@ -1,32 +1,29 @@
-import { rails as usage } from './modules/usage'
-import { rails } from './modules/rails'
+import { hedgesnear as usage } from './modules/usage'
+import { hedges } from './modules/hedges'
 import { ignoredBlocks, allowedBlocks } from './modules/blocks'
-import { getRadius, getRegion, transformIDs } from './modules/utils'
+import { transformIDs } from './modules/utils'
 
 importPackage(Packages.com.sk89q.worldedit)
 importPackage(Packages.com.sk89q.worldedit.math)
 importPackage(Packages.com.sk89q.worldedit.blocks)
 
-context.checkArgs(0, 1, usage)
-
-const region = getRegion()
+context.checkArgs(1, 2, usage)
 
 const options = {
-  block: 'iron_block',
-  height: 1,
-  offset: 0,
-  regex: '^.*$',
-  radius: getRadius(region),
-  center: region.center,
+  block: 'leaves:4',
+  height: 2,
+  offset: 1,
+  radius: Number.parseFloat(argv[1]),
+  center: player.getLocation(),
   onGround: true,
   ignoreBuildings: true,
   ignoreVegetation: true,
   ignoredBlocks,
   allowedBlocks,
-  ...JSON.parse(argv[1] || '{}')
+  ...JSON.parse(argv[2] || '{}')
 }
 
 transformIDs(options, 'ignoredBlocks')
 transformIDs(options, 'allowedBlocks')
 
-rails(options)
+hedges(options)